Fixed warnings about the numbers conversions

+++ SpeakHere/AudioViews/GLLevelMeter.m
@@ -239,7 +239,7 @@ Copyright (C) 2012 Apple Inc. All Rights Reserved.
 		if (_peakLevel > 0.)
 		{
 			peakLight = _peakLevel * _numLights;
-			if (peakLight >= _numLights) peakLight = _numLights - 1;
+			if (peakLight >= _numLights) peakLight = (int)(_numLights - 1);
 		}
 		
 		for (light_i=0; light_i<_numLights; light_i++)
